+++ Tests/KeyboardKitTests/_Keyboard/KeyboardContext+KeyboardTypeTests.swift
@@ -19,6 +19,7 @@ class KeyboardContext_KeyboardTypeTests: XCTestCase {
     override func setUp() {
         proxy = MockTextDocumentProxy()
         context = KeyboardContext()
+        context.isAutocapitalizationEnabled = true
         context.sync(with: MockKeyboardInputViewController())
         context.originalTextDocumentProxy = proxy
         context.keyboardType = .alphabetic(.lowercased)
@@ -43,7 +44,7 @@ class KeyboardContext_KeyboardTypeTests: XCTestCase {
 
 
     func testPreferredKeyboardTypeIgnoresAutoCapitalizationIfOverrideIsSet() {
-        context.isAutoCapitalizationEnabled = false
+        context.isAutocapitalizationEnabled = false
         let current = Keyboard.KeyboardType.alphabetic(.lowercased)
         let type = UITextAutocapitalizationType.allCharacters
         XCTAssertEqual(result(for: current, preCursorPart: "", type: type), current)